Reworked the connection between carriages and orders (added entity of pilot table). Changed type of field 'point' in tables carriages and routes to 'int'

+++ src/main/java/com/example/delivery_system/entity/RoutesForCarriage.java
@@ -19,7 +19,7 @@ public class RoutesForCarriage implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-    private byte point;
+    private int point;
 
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
